avanzatech_blog/requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.1.6
django-environ==0.12.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pluggy==1.5.0
psycog==0.0.1
psycopg==3.2.6
psycopg-pool==3.2.6
pytest==8.3.5
pytest-django==4.10.0
PyYAML==6.0.2
referencing==0.36.2
rest-framework-simplejwt==0.0.2
rpds-py==0.23.1
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
